requirements.txt
python-telegram-bot>=20.5
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiohttp>=3.8.5
openai>=1.3.4
psutil>=5.9.6
httpx>=0.24.1
translate==3.6.1 